requirements.txt
asgiref
Django
django-extensions
djangorestframework
graphviz
pydotplus
pyparsing
pytz
sqlparse
typing_extensions
gunicorn
